+++ tests/sequential/prometheus_metrics/prometheus_metrics_test.go
@@ -1544,8 +1544,11 @@ func testEmptyUpstreamResponse(t *testing.T, data templateData) {
 		for _, metric := range family.GetMetric() {
 			labels := metric.GetLabel()
 			if ExtractPrometheusLabelValue("namespace", labels) == testNamespace &&
-				ExtractPrometheusLabelValue("scaledObject", labels) == emptyUpstreamScaledObjectName &&
+				ExtractPrometheusLabelValue("scaledResource", labels) == emptyUpstreamScaledObjectName &&
 				ExtractPrometheusLabelValue("triggerName", labels) == "empty-upstream-trigger" &&
+				ExtractPrometheusLabelValue("metricName", labels) == "s0-prometheus" &&
+				ExtractPrometheusLabelValue("resourceType", labels) == "ScaledObject" &&
+				ExtractPrometheusLabelValue("ignoreNullValues", labels) == "false" &&
 				metric.GetCounter().GetValue() >= 1 {
 				return true
 			}